requirements.txt
pytest>=7.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0 